Add shared MemoryAdapter test helper exported from @haverstack/core/testing
Extracts the duplicated in-memory StackAdapter from stack.test.ts and
scoped-stack.test.ts into a single MemoryAdapter class with offset-based
cursor pagination, published via a dedicated testing entrypoint to avoid
polluting production bundles. Both test files are updated to import from
the new helper.

Closes #22
